Moved shared functions to new common.js file.
