+++ src/admin-events/admin-events.js
@@ -1,6 +1,6 @@
 const ethers = require('ethers');
-
 const { Finding, FindingSeverity, FindingType } = require('forta-agent');
+const { getAbi, filterAndParseLogs, extractEventArgs } = require('../common');
 
 // load any agent configuration parameters
 const config = require('../../agent-config.json');
@@ -18,25 +18,6 @@ function getEvents(contractName, adminEvents) {
     return {}; // no events for this contract
   }
   return events;
-}
-
-function getAbi(abiName) {
-  // eslint-disable-next-line global-require,import/no-dynamic-require
-  const { abi } = require(`../../abi/${abiName}`);
-  return abi;
-}
-
-// helper function that identifies key strings in the args array obtained from log parsing
-// these key-value pairs will be added to the metadata as event args
-// all values are converted to strings so that BigNumbers are readable
-function extractEventArgs(args) {
-  const eventArgs = {};
-  Object.keys(args).forEach((key) => {
-    if (Number.isNaN(Number(key))) {
-      eventArgs[key] = args[key].toString();
-    }
-  });
-  return eventArgs;
 }
 
 // helper function to create alerts
@@ -67,22 +48,6 @@ function createAlert(
       eventArgs,
     },
   });
-}
-
-// helper function to filter logs based on contract addresses and event names
-function filterAndParseLogs(logs, address, iface, eventNames) {
-  // collect logs only from the contracts of interest
-  const contractLogs = logs.filter((log) => log.address === address);
-  if (contractLogs.length === 0) {
-    return [];
-  }
-
-  // decode logs and filter on the ones we are interested in
-  const parse = (log) => iface.parseLog(log);
-  const filter = (log) => eventNames.indexOf(log.name) !== -1;
-  const parsedLogs = contractLogs.map(parse).filter(filter);
-
-  return parsedLogs;
 }
 
 function provideInitialize(data) {
