+++ src/common.js
@@ -1,0 +1,42 @@
+/* Library of common functions used by multiple handlers/agents */
+
+function getAbi(abiName) {
+  // eslint-disable-next-line global-require,import/no-dynamic-require
+  const { abi } = require(`../abi/${abiName}`);
+  return abi;
+}
+
+// helper function to filter logs based on contract addresses and event names
+function filterAndParseLogs(logs, address, iface, eventNames) {
+  // collect logs only from the contracts of interest
+  const contractLogs = logs.filter((log) => log.address === address);
+  if (contractLogs.length === 0) {
+    return [];
+  }
+
+  // decode logs and filter on the ones we are interested in
+  const parse = (log) => iface.parseLog(log);
+  const filter = (log) => eventNames.indexOf(log.name) !== -1;
+  const parsedLogs = contractLogs.map(parse).filter(filter);
+
+  return parsedLogs;
+}
+
+// helper function that identifies key strings in the args array obtained from log parsing
+// these key-value pairs will be added to the metadata as event args
+// all values are converted to strings so that BigNumbers are readable
+function extractEventArgs(args) {
+  const eventArgs = {};
+  Object.keys(args).forEach((key) => {
+    if (Number.isNaN(Number(key))) {
+      eventArgs[key] = args[key].toString();
+    }
+  });
+  return eventArgs;
+}
+
+module.exports = {
+  getAbi,
+  filterAndParseLogs,
+  extractEventArgs,
+};
